Features/fix audit issues part 2

+++ contracts/Migrations.sol
@@ -2,22 +2,22 @@ pragma solidity 0.5.10;
 
 contract Migrations {
     address public owner;
-  uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
-  constructor() public {
-    owner = msg.sender;
-  }
+    constructor() public {
+        owner = msg.sender;
+    }
 
-  modifier restricted() {
-    if (msg.sender == owner) _;
-  }
+    modifier restricted() {
+        if (msg.sender == owner) _;
+    }
 
-  function setCompleted(uint completed) external restricted {
-    lastCompletedMigration = completed;
-  }
+    function setCompleted(uint256 completed) external restricted {
+        lastCompletedMigration = completed;
+    }
 
-  function upgrade(address newAddress) external restricted {
-    Migrations upgraded = Migrations(newAddress);
-    upgraded.setCompleted(lastCompletedMigration);
-  }
+    function upgrade(address newAddress) external restricted {
+        Migrations upgraded = Migrations(newAddress);
+        upgraded.setCompleted(lastCompletedMigration);
+    }
 }
